updated CURD

+++ src/main/java/com/t13g2/forum/logic/commands/DeleteThreadCommand.java
@@ -1,20 +1,21 @@
 package com.t13g2.forum.logic.commands;
 
+import static com.t13g2.forum.commons.core.Messages.MESSAGE_INVALID_THREAD_ID;
+import static com.t13g2.forum.commons.core.Messages.MESSAGE_NOT_LOGIN;
+import static com.t13g2.forum.commons.core.Messages.MESSAGE_NOT_THREAD_OWNER;
 import static com.t13g2.forum.logic.parser.CliSyntax.PREFIX_THREAD_ID;
 import static java.util.Objects.requireNonNull;
 
 import com.t13g2.forum.logic.CommandHistory;
 import com.t13g2.forum.logic.commands.exceptions.CommandException;
+import com.t13g2.forum.model.Context;
 import com.t13g2.forum.model.Model;
 import com.t13g2.forum.model.UnitOfWork;
+import com.t13g2.forum.storage.forum.EntityDoesNotExistException;
 
 /**
  * Delete a certain thread. Only admin could delete threads from others,
- * user could only delete threads created by himself/herself.
- */
-
-/**
- * Example: deleteThread tId/1
+ * user could only delete threads created by his/her own.
  */
 public class DeleteThreadCommand extends Command {
     public static final String COMMAND_WORD = "deleteThread";
@@ -26,28 +27,37 @@ public class DeleteThreadCommand extends Command {
         + PREFIX_THREAD_ID + "1";
 
     public static final String MESSAGE_SUCCESS = "Thread deleted: %1$s";
-    public static final String MESSAGE_INVALID_THREAD_ID = "Invalid Thread id";
-
     private final int threadId;
 
     /**
      * Creates an DeleteThreadCommand to delete the specified {@code ForumThread}
      */
     public DeleteThreadCommand(int threadId) {
+        requireNonNull(threadId);
         this.threadId = threadId;
     }
 
     @Override
     public CommandResult execute(Model model, CommandHistory history) throws CommandException {
         requireNonNull(model);
+        if (!Context.getInstance().isLoggedIn()) {
+            throw new CommandException(MESSAGE_NOT_LOGIN);
+        }
         try (UnitOfWork unitOfWork = new UnitOfWork()) {
+            if (Context.getInstance().getCurrentUser().getId()
+                    != unitOfWork.getForumThreadRepository().getThread(threadId).getCreatedByUserId()) {
+                throw new CommandException(MESSAGE_NOT_THREAD_OWNER);
+            }
             //delete the thread according to the threadId from the memory repository
             unitOfWork.getForumThreadRepository().deleteThread(threadId);
             //update to local database
             unitOfWork.commit();
+        } catch (EntityDoesNotExistException e) {
+            throw new CommandException(MESSAGE_INVALID_THREAD_ID);
+        } catch (CommandException e) {
+            throw e;
         } catch (Exception e) {
             e.printStackTrace();
-            throw new CommandException(MESSAGE_INVALID_THREAD_ID);
         }
         String message = "\n"
             + "Thread ID: " + threadId + "\n";
